refactor: Update asset and category retrieval methods to use Long instead of Integer; enhance DetailItemController with new alert confirmation and asset handling features; remove unused alert message class and consolidate styles into partials

+++ src/main/java/io/github/codenilson/smartpat/application/usecase/asset/GetAssetById.java
@@ -14,7 +14,7 @@ public class GetAssetById {
         this.assetRepository = assetRepository;
     }
 
-    public Asset execute(Integer assetId) {
+    public Asset execute(Long assetId) {
         return assetRepository.findById(assetId);
     }
 }
